requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
langchain_core